refactor line into smaller files; only show leap motion instructions when leap is detected
