+++ src/custom-typings/leapjs.d.ts
@@ -1,6 +1,6 @@
 // most of this is taken from https://github.com/logotype/LeapMotionTS/blob/master/build/leapmotionts-2.2.4.d.ts
 declare module 'leapjs' {
-    export function loop(cb: (frame: Frame) => void): void;
+    export function loop(cb: (frame: Frame) => void): Controller;
     /**
      * The EventDispatcher class provides strongly typed events.
      */
@@ -250,7 +250,7 @@ declare module 'leapjs' {
          *
          * @param a A number.
          * @param b A number.
-         * @param coefficient The interpolation coefficient [0-1].
+     * @param coefficient The interpolation coefficient [0-1].
          * @return The interpolated number.
          *
          */
@@ -324,16 +324,9 @@ declare module 'leapjs' {
          * Socket connection.
          */
         connection: WebSocket;
-        /**
-         * @private
-         * Reports whether this Controller is connected to the Leap Motion Controller.
-         */
-        _isConnected: boolean;
-        /**
-         * @private
-         * Reports whether gestures is enabled.
-         */
-        _isGesturesEnabled: boolean;
+        lastConnectionFrame: Frame;
+        lastFrame: Frame;
+        lastValidFrame: Frame;
         /**
          * Constructs a Controller object.
          * @param host IP or hostname of the computer running the Leap software.
@@ -414,19 +407,19 @@ declare module 'leapjs' {
         /**
          * Reports whether this Controller is connected to the Leap Motion Controller.
          *
-         * <p>When you first create a Controller object, <code>isConnected()</code> returns false.
+         * <p>When you first create a Controller object, <code>connected()</code> returns false.
          * After the controller finishes initializing and connects to
-         * the Leap, <code>isConnected()</code> will return true.</p>
+         * the Leap, <code>connected()</code> will return true.</p>
          *
          * <p>You can either handle the onConnect event using a event listener
-         * or poll the <code>isConnected()</code> if you need to wait for your
+         * or poll the <code>connected()</code> if you need to wait for your
          * application to be connected to the Leap before performing
          * some other operation.</p>
          *
          * @return True, if connected; false otherwise.
          *
          */
-        isConnected(): boolean;
+        connected(): boolean;
     }
     /**
      * The InteractionBox class represents a box-shaped region completely within
@@ -1388,6 +1381,21 @@ declare module 'leapjs' {
          */
         controller: Controller;
         /**
+         * Reports whether this Frame instance is valid.
+         *
+         * <p>A valid Frame is one generated by the LeapMotion object that contains
+         * tracking data for all detected entities. An invalid Frame contains
+         * no actual tracking data, but you can call its functions without risk
+         * of a null pointer exception. The invalid Frame mechanism makes it
+         * more convenient to track individual data across the frame history.</p>
+         *
+         * <p>For example, you can invoke: <code>var finger:Finger = leap.frame(n).finger(fingerID);</code>
+         * for an arbitrary Frame history value, "n", without first checking whether
+         * frame(n) returned a null object.<br/>
+         * (You should still check that the returned Finger instance is valid.)</p>
+         */
+        valid: boolean;
+        /**
          * Constructs a Frame object.
          *
          * <p>Frame instances created with this constructor are invalid.
@@ -1620,24 +1628,6 @@ declare module 'leapjs' {
          *
          */
         isEqualTo(other: Frame): boolean;
-        /**
-         * Reports whether this Frame instance is valid.
-         *
-         * <p>A valid Frame is one generated by the LeapMotion object that contains
-         * tracking data for all detected entities. An invalid Frame contains
-         * no actual tracking data, but you can call its functions without risk
-         * of a null pointer exception. The invalid Frame mechanism makes it
-         * more convenient to track individual data across the frame history.</p>
-         *
-         * <p>For example, you can invoke: <code>var finger:Finger = leap.frame(n).finger(fingerID);</code>
-         * for an arbitrary Frame history value, "n", without first checking whether
-         * frame(n) returned a null object.<br/>
-         * (You should still check that the returned Finger instance is valid.)</p>
-         *
-         * @return True, if this is a valid Frame object; false otherwise.
-         *
-         */
-        isValid(): boolean;
         /**
          * Returns an invalid Frame object.
          *
