+++ src/sketches/line/index.tsx
@@ -1,0 +1,3 @@
+import { LineSketch } from "./line";
+
+export default LineSketch;
