+++ src/sketches/line/leapMotion.ts
@@ -1,0 +1,100 @@
+import * as Leap from "leapjs";
+import * as THREE from "three";
+
+import { map } from "../../math/index";
+import { Attractor, HandMesh } from "./attractor";
+import { attractors, canvas, globalFrame, instructionsEl, mouseX, mouseY, scene, setMousePosition } from "./line";
+
+const boneGeometry = new THREE.SphereGeometry(10, 3, 3);
+const boneMaterial = new THREE.LineBasicMaterial({
+    // color: 0xefeffb,
+    color: 0xadd6b6,
+    linewidth: 5,
+    transparent: true,
+    opacity: 1,
+});
+
+export function initLeap() {
+    const controller = Leap.loop((frame: Leap.Frame) => {
+        if (frame.hands.length > 0) {
+            instructionsEl.setLastRenderedFrame(globalFrame);
+        }
+        attractors.forEach((attractor) => {
+            if (attractor.handMesh != null) {
+                attractor.handMesh.visible = false;
+            }
+            attractor.mesh.visible = false;
+            attractor.power = 0;
+        });
+        frame.hands.filter((hand) => hand.valid).forEach((hand, index) => {
+            const position = hand.indexFinger.bones[3].center();
+
+            const {x, y} = mapLeapToThreePosition(position);
+            setMousePosition(x, y);
+
+            const attractor = attractors[index];
+            attractor.x = x;
+            attractor.y = y;
+            attractor.mesh.position.x = x;
+            attractor.mesh.position.y = y;
+
+            attractor.mesh.visible = true;
+            if (hand.indexFinger.extended) {
+                // position[2] goes from -300 to 300
+                const wantedPower = Math.pow(7, (-position[2] + 350) / 200);
+                attractor.power = attractor.power * 0.5 + wantedPower * 0.5;
+            } else {
+                attractor.power = attractor.power * 0.5;
+            }
+
+            updateHandMesh(attractor, hand);
+            attractor.handMesh!.visible = true;
+        });
+    });
+    return controller;
+}
+
+function updateHandMesh(attractor: Attractor, hand: Leap.Hand) {
+    if (attractor.handMesh == null) {
+        attractor.handMesh = new THREE.Object3D() as HandMesh;
+        scene.add(attractor.handMesh);
+    }
+    const handMesh = attractor.handMesh;
+    hand.fingers.forEach((finger) => {
+        if (handMesh["finger" + finger.type] == null) {
+            const fingerLine = new THREE.Line(new THREE.Geometry(), boneMaterial);
+            handMesh["finger" + finger.type] = fingerLine;
+            handMesh.add(fingerLine);
+        }
+        const fingerGeometry = handMesh["finger" + finger.type].geometry as THREE.Geometry;
+        finger.bones.forEach((bone) => {
+            // create sphere for every bone
+            const id = finger.type + ',' + bone.type;
+            if (handMesh[id] == null) {
+                const boneMesh = new THREE.Mesh(boneGeometry, boneMaterial);
+                handMesh[id] = boneMesh;
+                handMesh.add(boneMesh);
+            }
+            const position = mapLeapToThreePosition(bone.center());
+            handMesh[id].position.copy(position);
+
+            // create a line for every finger
+            if (fingerGeometry.vertices[bone.type] == null) {
+                fingerGeometry.vertices.push(new THREE.Vector3());
+            }
+            fingerGeometry.vertices[bone.type].copy(position);
+            fingerGeometry.verticesNeedUpdate = true;
+        });
+    });
+}
+
+function mapLeapToThreePosition(position: number[]) {
+    const range = [0.2, 0.8];
+    // position[0] is left/right; left is negative, right is positive. each unit is one millimeter
+    const x = map(position[0], -200, 200, canvas.width * range[0],  canvas.width * range[1]);
+    // 40 is about 4cm, 1 inch, to 35cm = 13 inches above
+    const y = map(position[1], 350, 40,   canvas.height * range[0], canvas.height * range[1]);
+    // put the leap stuff close to the camera so the hand is always visible
+    const z = 300;
+    return new THREE.Vector3(x, y, z);
+}
